+++ src/main/java/eu/trentorise/opendata/commons/exceptions/TodNotFoundException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package exceptions;
+package eu.trentorise.opendata.commons.exceptions;
 
 /**
  * A runtime exception to raise when something is not found.