+++ src/main/java/eu/trentorise/opendata/commons/exceptions/TodParseException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package eu.trentorise.opendata.commons.exceptions;
 
 public class TodParseException extends TodException {